# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "kenaz-ml",
    "description": "kenaz-ml \u2014 the ML sidecar for Sigil (local mode)",
    "version": "0.1.0"
  },
  "paths": {
...
    "/fleet/train/{model_name}": {
      "post": {
        "summary": "Fleet Train",
        "description": "Train a fleet model for a specific team.",
        "operationId": "fleet_train_fleet_train__model_name__post",
        "parameters": [
          {
            "name": "model_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FleetTrainRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FleetTrainResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/fleet/predict/{model_name}": {
      "get": {
        "summary": "Fleet Predict",
        "description": "Get predictions from a trained fleet model.",
        "operationId": "fleet_predict_fleet_predict__model_name__get",
        "parameters": [
          {
            "name": "model_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Name"
            }
          },
          {
            "name": "team_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Team Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FleetPredictResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FleetPredictResponse": {
        "properties": {
          "model": {
            "type": "string",
            "title": "Model"
          },
          "team_id": {
            "type": "integer",
            "title": "Team Id"
          },
          "predictions": {
            "additionalProperties": true,
            "type": "object",
            "title": "Predictions"
          },
          "trained_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trained At"
          },
          "samples": {
            "type": "integer",
            "title": "Samples",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "model",
          "team_id",
          "predictions"
        ],
        "title": "FleetPredictResponse"
      },
      "FleetTrainRequest": {
        "properties": {
          "team_id": {
            "type": "integer",
            "title": "Team Id"
          },
          "data": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "minItems": 5,
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "team_id",
          "data"
        ],
        "title": "FleetTrainRequest"
      },
      "FleetTrainResponse": {
        "properties": {
          "model": {
            "type": "string",
            "title": "Model"
          },
          "team_id": {
            "type": "integer",
            "title": "Team Id"
          },
          "samples": {
            "type": "integer",
            "title": "Samples"
          },
          "trained_at": {
            "type": "string",
            "title": "Trained At"
          },
          "metrics": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Metrics"
          }
        },
        "type": "object",
        "required": [
          "model",
          "team_id",
          "samples",
          "trained_at",
          "metrics"
        ],
        "title": "FleetTrainResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}